+++ internal/topic/topicwriterinternal/queue_test.go
@@ -14,7 +14,6 @@ import (
 
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/empty"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/grpcwrapper/rawtopic/rawtopicwriter"
-	"github.com/ydb-platform/ydb-go-sdk/v3/internal/xatomic"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/xcontext"
 )
 
@@ -142,7 +141,7 @@ func TestMessageQueue_GetMessages(t *testing.T) {
 		}()
 
 		readFinished := make(empty.Chan)
-		var lastReadSeqNo xatomic.Int64
+		var lastReadSeqNo atomic.Int64
 
 		readCtx, readCancel := xcontext.WithCancel(ctx)
 		defer readCancel()
@@ -406,6 +405,26 @@ func TestQueuePanicOnOverflow(t *testing.T) {
 		q.lastWrittenIndex = maxInt
 		q.addMessageNeedLock(messageWithDataContent{})
 	})
+}
+
+func TestRegressionIssue1038_ReceiveAckAfterCloseQueue(t *testing.T) {
+	counter := 0
+
+	q := newMessageQueue()
+	q.OnAckReceived = func(count int) {
+		counter -= count
+	}
+	require.NoError(t, q.AddMessages(newTestMessagesWithContent(1)))
+	counter++
+
+	require.NoError(t, q.Close(errors.New("test err")))
+	require.ErrorIs(t, q.AcksReceived([]rawtopicwriter.WriteAck{
+		{
+			SeqNo:              1,
+			MessageWriteStatus: rawtopicwriter.MessageWriteStatus{},
+		},
+	}), errAckOnClosedMessageQueue)
+	require.Zero(t, counter)
 }
 
 func TestQueue_Ack(t *testing.T) {