+++ table/table.go
@@ -2,20 +2,17 @@ package table
 
 import (
 	"context"
-	"sort"
 	"time"
 
-	"github.com/ydb-platform/ydb-go-genproto/protos/Ydb"
 	"github.com/ydb-platform/ydb-go-genproto/protos/Ydb_Table"
 
-	"github.com/ydb-platform/ydb-go-sdk/v3/internal/allocator"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/closer"
+	"github.com/ydb-platform/ydb-go-sdk/v3/internal/params"
+	"github.com/ydb-platform/ydb-go-sdk/v3/internal/types"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/value"
-	"github.com/ydb-platform/ydb-go-sdk/v3/internal/xstring"
 	"github.com/ydb-platform/ydb-go-sdk/v3/retry"
 	"github.com/ydb-platform/ydb-go-sdk/v3/table/options"
 	"github.com/ydb-platform/ydb-go-sdk/v3/table/result"
-	"github.com/ydb-platform/ydb-go-sdk/v3/table/types"
 	"github.com/ydb-platform/ydb-go-sdk/v3/trace"
 )
 
@@ -144,7 +141,7 @@ type Session interface {
 		ctx context.Context,
 		tx *TransactionControl,
 		query string,
-		params *QueryParameters,
+		params *params.Parameters,
 		opts ...options.ExecuteDataQueryOption,
 	) (txr Transaction, r result.Result, err error)
 
@@ -167,21 +164,21 @@ type Session interface {
 	StreamExecuteScanQuery(
 		ctx context.Context,
 		query string,
-		params *QueryParameters,
+		params *params.Parameters,
 		opts ...options.ExecuteScanQueryOption,
 	) (_ result.StreamResult, err error)
 
 	BulkUpsert(
 		ctx context.Context,
 		table string,
-		rows types.Value,
+		rows value.Value,
 		opts ...options.BulkUpsertOption,
 	) (err error)
 
 	ReadRows(
 		ctx context.Context,
 		path string,
-		keys types.Value,
+		keys value.Value,
 		opts ...options.ReadRowsOption,
 	) (_ result.Result, err error)
 
@@ -243,13 +240,13 @@ type TransactionActor interface {
 	Execute(
 		ctx context.Context,
 		query string,
-		params *QueryParameters,
+		params *params.Parameters,
 		opts ...options.ExecuteDataQueryOption,
 	) (result.Result, error)
 	ExecuteStatement(
 		ctx context.Context,
 		stmt Statement,
-		params *QueryParameters,
+		params *params.Parameters,
 		opts ...options.ExecuteDataQueryOption,
 	) (result.Result, error)
 }
@@ -270,7 +267,7 @@ type Statement interface {
 	Execute(
 		ctx context.Context,
 		tx *TransactionControl,
-		params *QueryParameters,
+		params *params.Parameters,
 		opts ...options.ExecuteDataQueryOption,
 	) (txr Transaction, r result.Result, err error)
 	NumInput() int
@@ -455,113 +452,20 @@ func SnapshotReadOnlyTxControl() *TransactionControl {
 
 // QueryParameters
 type (
-	queryParams     map[string]types.Value
-	ParameterOption interface {
-		Name() string
-		Value() types.Value
-	}
-	parameterOption struct {
-		name  string
-		value types.Value
-	}
-	QueryParameters struct {
-		m queryParams
-	}
+	ParameterOption = params.NamedValue
+	QueryParameters = params.Parameters
 )
 
-func (p parameterOption) Name() string {
-	return p.name
-}
-
-func (p parameterOption) Value() types.Value {
-	return p.value
-}
-
-func (qp queryParams) ToYDB(a *allocator.Allocator) map[string]*Ydb.TypedValue {
-	if qp == nil {
-		return nil
-	}
-	params := make(map[string]*Ydb.TypedValue, len(qp))
-	for k, v := range qp {
-		params[k] = value.ToYDB(v, a)
-	}
-
-	return params
-}
-
-func (q *QueryParameters) Params() queryParams {
-	if q == nil {
-		return nil
-	}
-
-	return q.m
-}
-
-func (q *QueryParameters) Count() int {
-	if q == nil {
-		return 0
-	}
-
-	return len(q.m)
-}
-
-func (q *QueryParameters) Each(it func(name string, v types.Value)) {
-	if q == nil {
-		return
-	}
-	for key, v := range q.m {
-		it(key, v)
-	}
-}
-
-func (q *QueryParameters) names() []string {
-	if q == nil {
-		return nil
-	}
-	names := make([]string, 0, len(q.m))
-	for k := range q.m {
-		names = append(names, k)
-	}
-	sort.Strings(names)
-
-	return names
-}
-
-func (q *QueryParameters) String() string {
-	buffer := xstring.Buffer()
-	defer buffer.Free()
-
-	buffer.WriteByte('{')
-	for i, name := range q.names() {
-		if i != 0 {
-			buffer.WriteByte(',')
-		}
-		buffer.WriteByte('"')
-		buffer.WriteString(name)
-		buffer.WriteString("\":")
-		buffer.WriteString(q.m[name].Yql())
-	}
-	buffer.WriteByte('}')
-
-	return buffer.String()
-}
-
 func NewQueryParameters(opts ...ParameterOption) *QueryParameters {
-	q := &QueryParameters{
-		m: make(queryParams, len(opts)),
+	qp := QueryParameters(make([]*params.Parameter, len(opts)))
+	for i, opt := range opts {
+		qp[i] = params.Named(opt.Name(), opt.Value())
 	}
-	q.Add(opts...)
 
-	return q
+	return &qp
 }
 
-func (q *QueryParameters) Add(params ...ParameterOption) {
-	for _, param := range params {
-		q.m[param.Name()] = param.Value()
-	}
-}
-
-func ValueParam(name string, v types.Value) ParameterOption {
+func ValueParam(name string, v value.Value) ParameterOption {
 	switch len(name) {
 	case 0:
 		panic("empty name")
@@ -571,10 +475,7 @@ func ValueParam(name string, v types.Value) ParameterOption {
 		}
 	}
 
-	return &parameterOption{
-		name:  name,
-		value: v,
-	}
+	return params.Named(name, v)
 }
 
 type Options struct {